requirements.txt
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
black==21.5b0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
dj-database-url==0.5.0
Django==3.2.2
django-heroku==0.3.1
django-stdimage==5.3.0
django-storages==1.11.1
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==2.10
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
Pillow==8.2.0
protobuf==3.16.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
sentry-sdk==1.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
